requirements.txt
playwright==1.57.0
requests==2.32.5
python-dotenv==1.2.1
tqdm==4.67.1
pillow